% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radii.R
\name{add_radii_of_gyration_unc}
\alias{add_radii_of_gyration_unc}
\title{Add radii of gyration uncertainties}
\usage{
add_radii_of_gyration_unc(df)
}
\arguments{
\item{df}{A data frame with (at least) these columns: \code{id}, \code{mass}, \code{Cx},
\code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy}, \code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint},
\code{sigma_mass}, \code{sigma_Cx}, \code{sigma_Cy}, \code{sigma_Cz},\code{sigma_Ixy}, \code{sigma_Ixz},
\code{sigma_Iyz}.}
}
\value{
A data frame with the same columns as \code{df}, plus
radii of gyration in columns \code{sigma_kx}, \code{sigma_ky}, and \code{sigma_kz}.`
}
\description{
\code{add_radii_of_gyration_unc()} adds calculated radii of gyration uncertainties to a data frame
of rolled-up mass properties and uncertainties.

Radii of gyration uncertainties are calculated directly from moments of inertia and mass
and their uncertainties;
they are not recursively-defined, and do not require a rollup method.
}
\examples{
sawe_table_rollup <- rollup_mass_props(sawe_tree, sawe_table)
add_radii_of_gyration_unc(add_radii_of_gyration(sawe_table_rollup))
}
