% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{combine_mass_props}
\alias{combine_mass_props}
\title{Combine mass properties}
\usage{
combine_mass_props(vl)
}
\arguments{
\item{vl}{list of mass properties lists}
}
\value{
Combined mass properties list
}
\description{
Combine mass properties
}
\examples{
leaves <- names(igraph::neighbors(test_tree, "A.3", mode = "in"))
vl <- Map(f = function(id) get_mass_props(test_table, id), leaves)
combine_mass_props(vl)
}
