% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{combine_mass_props_and_unc}
\alias{combine_mass_props_and_unc}
\title{Combine mass properties and uncertainties}
\usage{
combine_mass_props_and_unc(vl)
}
\arguments{
\item{vl}{list of mass properties and uncertainties lists}
}
\value{
combined mass properties and uncertainties
}
\description{
\code{combine_mass_props_and_unc()} is a convenience wrapper that concatentates the
results of \code{combine_mass_props()} and \code{combine_mass_props_unc()}.
}
\examples{
leaves <- names(igraph::neighbors(sawe_tree, "Combined", mode = "in"))
vl <- Map(f = function(id) get_mass_props_and_unc(sawe_table, id), leaves)
combine_mass_props_and_unc(vl)
}
