% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{combine_mass_props_unc}
\alias{combine_mass_props_unc}
\title{Combine mass properties uncertainties}
\usage{
combine_mass_props_unc(vl, r)
}
\arguments{
\item{vl}{list of mass properties and uncertainties lists}

\item{r}{mass properties previously combined}
}
\value{
combined mass properties and uncertainties
}
\description{
Combine mass properties uncertainties
}
\examples{
leaves <- names(igraph::neighbors(sawe_tree, "Combined", mode = "in"))
vl <- Map(f = function(id) get_mass_props_and_unc(sawe_table, id), leaves)
combine_mass_props_unc(vl, r = get_mass_props(sawe_table, "Combined"))
}
