% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radii.R
\name{combine_radii_of_gyration_unc}
\alias{combine_radii_of_gyration_unc}
\title{Combine radii of gyration uncertainties}
\usage{
combine_radii_of_gyration_unc(mpl, amp)
}
\arguments{
\item{mpl}{A list of mass properties and uncertainties lists, each of which contains
the following named elements:
\itemize{
\item \code{mass} Numeric mass.
\item \code{center_mass} Numeric 3-vector center of mass.
\item \code{sigma_mass} mass uncertainty
\item \code{sigma_center_mass} center of mass uncertainty (3-dimensional numeric)
\item \code{radii_gyration} Numeric 3-vector radii of gyration.
}}

\item{amp}{A named list of mass properties for the aggregate containing the following
named elements:
\itemize{
\item \code{mass} Numeric mass.
\item \code{center_mass} Numeric 3-vector center of mass.
\item \code{radii_gyration} Numeric 3-vector radii of gyration.
}}
}
\value{
The radii of gyration uncertainties of the aggregate. A list with the following named elements:
\itemize{
\item \code{sigma_radii_gyration} Numeric 3-vector radii of gyration uncertainties.
}
}
\description{
\code{combine_radii_of_gyration_unc()} calculates the radii of gyration uncertainties of an aggregate from
the mass properties and uncertainties and radii of gyration of its constituents and the mass properties
and radii of gyration of the aggregate.
}
\details{
See vignette("massProps", package = "massProps") for details on the algorithms
employed.
}
