% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{df_set_mass_props}
\alias{df_set_mass_props}
\title{Set mass properties for a row in a data frame}
\usage{
df_set_mass_props(df, id, v)
}
\arguments{
\item{df}{A data frame}

\item{id}{ID value of the desired row}

\item{v}{#' A list with the following named elements:
\itemize{
\item \code{mass} mass (numeric)
\item \code{center_mass} center of mass (3-dimensional numeric)
\item \code{inertia} Inertia tensor (3x3 numeric matrix)
\item \code{point} Logical indicating point mass, i.e., negligible inertia
\item \code{poi_conv} Enumeration c("+", "-") indicating sign convention for products of inertia
}}
}
\value{
The updated data frame
}
\description{
\code{df_set_mass_props()} sets mass properties for a specified row in a data frame
with (at least) these columns: \code{mass}, \code{Cx}, \code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy},
\code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}.
}
