% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{df_set_mass_props_and_unc}
\alias{df_set_mass_props_and_unc}
\title{Set mass properties and uncertainties for a row in a data frame}
\usage{
df_set_mass_props_and_unc(df, id, v)
}
\arguments{
\item{df}{A data frame}

\item{id}{ID value of the desired row}

\item{v}{#' A list with the following named elements:
\itemize{
\item \code{mass} mass (numeric)
\item \code{center_mass} center of mass (3-dimensional numeric)
\item \code{inertia} Inertia tensor (3x3 numeric matrix)
\item \code{point} Logical indicating point mass, i.e., negligible inertia
\item \code{poi_conv} Enumeration c("+", "-") indicating sign convention for products of inertia
\item \code{σ_mass} mass uncertainty (numeric)
\item \code{σ_center_mass} center of mass uncertainty (3-dimensional numeric)
\item \code{σ_inertia} Inertia tensor uncertainty (3x3 numeric matrix)
}}
}
\value{
The updated data frame
}
\description{
\code{df_set_mass_props_and_unc()} sets mass properties and uncertainties for a
specified row in a data frame with an \code{id} column.
}
\examples{
df <- data.frame(id = c("C.1.2.2.3.1.2.3", "C.1.2.2.3.2.1.1"))
v <- df_get_mass_props_and_unc(mp_table, "C.1.2.2.3.2.1.1")
v$poi_conv = "+"
df_set_mass_props_and_unc(df, "C.1.2.2.3.2.1.1", v)
}
