% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massProps-package.R
\docType{package}
\name{massProps-package}
\alias{massProps}
\alias{massProps-package}
\title{massProps: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
