% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tree.R
\docType{data}
\name{mp_tree}
\alias{mp_tree}
\title{Example Mass Properties Tree}
\format{
An igraph tree with edges from child id to parent id.
}
\usage{
mp_tree
}
\description{
Example Mass Properties Tree
}
\keyword{datasets}
