% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tree_small.R
\docType{data}
\name{mp_tree_small}
\alias{mp_tree_small}
\title{Example Small Mass Properties Tree}
\format{
An 'igraph' tree whose vertices are named as the values of the \code{id}
column of a mass properties table and whose directed edges point from child
id to parent id.
}
\usage{
mp_tree_small
}
\description{
Example Small Mass Properties Tree
}
\keyword{datasets}
