% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{rollup_mass_props}
\alias{rollup_mass_props}
\title{Roll up mass properties}
\usage{
rollup_mass_props(tree, ds, validate_ds = validate_mass_props_table, ...)
}
\arguments{
\item{tree}{An 'igraph' tree whose vertices are named as the values of the \code{id}
column of \code{ds} and whose directed edges point from child id to parent id.}

\item{ds}{A data set  with (at least) these properties: \code{id}, \code{mass}, \code{Cx},
\code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy}, \code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}.}

\item{validate_ds}{A validator for the tree and table, default \code{validate_mass_props_table()}}

\item{...}{Other parameters passed to \code{rollupTree::rollup()}}
}
\value{
The updated data set.
}
\description{
'rollup_mass_props()' rolls up mass properties in a data set such that the mass properties of each
non-leaf vertex item is the aggregation of those of its child items.
}
\examples{
rollup_mass_props(mp_tree_small, mp_table_small)

}
