% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radii.R
\name{rollup_radii_of_gyration_unc}
\alias{rollup_radii_of_gyration_unc}
\title{Roll up radii of gyration uncertainties}
\usage{
rollup_radii_of_gyration_unc(
  tree,
  ds,
  get_mpu = get_mass_props_and_unc,
  set_radii_unc = set_radii_of_gyration_unc,
  validate_ds = validate_mass_props_and_unc_table
)
}
\arguments{
\item{tree}{An 'igraph' tree whose vertices are named as the values of the \code{id}
column of \code{ds} and whose directed edges point from child id to parent id.}

\item{ds}{A data set  with (at least) these properties: \code{id}, \code{mass}, \code{Cx},
\code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy}, \code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}.}

\item{get_mpu}{A function to get mass properties and uncertainties for an item in a data set, called as
get_mpu(ds, id). Default: get_mass_props_and_unc().}

\item{set_radii_unc}{A function to set radii of gyration uncertainties for an item in a data set, called
as set_radii_unc(ds, target, mp). Default: set_radii_of_gyration_unc().}

\item{validate_ds}{A validator for the tree and table, default \code{validate_mass_props_and_unc_table()}}
}
\value{
A data set with the same properties as \code{ds}, plus
radii of gyration properties \code{sigma_kx}, \code{sigma_ky}, and \code{sigma_kz}.`
}
\description{
\code{rollup_radii_of_gyration_unc()} adds calculated radii of gyration uncertainties to a data set
of rolled-up mass properties and uncertainties.

Radii of gyration uncertainties are calculated directly from moments of inertia and mass
and their uncertainties; they are not recursively-defined. Radii of gyration uncertainties
for composite elements depend on uncertainties of their component elements.
}
\examples{
sawe_table_rollup <- rollup_mass_props(sawe_tree, sawe_table)
rollup_radii_of_gyration_unc(sawe_tree, add_radii_of_gyration(sawe_table_rollup))

}
