% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sawe_tree.R
\docType{data}
\name{sawe_tree}
\alias{sawe_tree}
\title{Example Mass Properties Tree}
\format{
An igraph tree with edges from child id to parent id.
}
\usage{
sawe_tree
}
\description{
Example Mass Properties Tree
}
\keyword{datasets}
