% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{set_mass_props_and_unc}
\alias{set_mass_props_and_unc}
\title{Set mass properties and uncertainties for an item in a data set}
\usage{
set_mass_props_and_unc(ds, id, mpu, set_by_id = df_set_by_id)
}
\arguments{
\item{ds}{A data set with an \code{id} property.}

\item{id}{The \code{id} value of the desired item}

\item{mpu}{A list containing the following named elements:
\itemize{
\item \code{mass} Numeric mass.
\item \code{center_mass} Numeric 3-vector center of mass.
\item \code{point} Logical indicating point mass. The inertia of point masses is excluded from calculations.
\item \code{poi_conv} Enumeration c("+", "-") indicating sign convention for products of inertia.
\item \code{inertia} Numeric 3x3 matrix inertia tensor. The signs of the products of inertia
are determined by \code{POIconv}. For example, \code{Ixy} is the \eqn{xy} element of the inertia
tensor if \code{POIconv} is "-"; it is the additive inverse of that value if \code{POIconv} is "+".
\item \code{sigma_mass} Numeric mass uncertainty.
\item \code{sigma_center_mass} Numeric 3-vector center of mass uncertainties.
\item \code{sigma_inertia} Numeric 3x3 matrix inertia tensor uncertainties.
}}

\item{set_by_id}{A function to set the value of a property for the selected item,
called as set_by_id(ds, id, property, value); default df_set_by_id()}
}
\value{
The updated data set.
}
\description{
\code{set_mass_props_and_unc()} is a convenience wrapper that combines the results of
\code{set_mass_props()} and \code{set_mass_props_unc()}.
}
\examples{
mpu <- c(get_mass_props_and_unc(sawe_table, "Widget"), poi_conv = "+")
set_mass_props_and_unc(sawe_table, "Combined", mpu)

}
