% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{set_mass_props_unc}
\alias{set_mass_props_unc}
\title{Set mass properties uncertainties for an item in a data set}
\usage{
set_mass_props_unc(ds, id, mpu, set_by_id = df_set_by_id)
}
\arguments{
\item{ds}{A data set with an \code{id} property.}

\item{id}{The \code{id} value of the desired item}

\item{mpu}{A list with the following named elements:
\itemize{
\item \code{sigma_mass} Numeric mass uncertainty.
\item \code{sigma_center_mass} Numeric 3-vector center of mass uncertainties.
\item \code{sigma_inertia} Numeric 3x3 matrix inertia tensor uncertainties.
}}

\item{set_by_id}{A function to set the value of a property for the selected item,
called as set_by_id(ds, id, property, value); default df_set_by_id()}
}
\value{
The updated data set.
}
\description{
\code{set_mass_props_unc()} sets mass properties uncertainties for a
selected item in a data set
}
\examples{
set_mass_props_unc(sawe_table, "Combined", get_mass_props_unc(sawe_table, "Widget"))

}
