% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{set_poi_conv_minus}
\alias{set_poi_conv_minus}
\title{Set POI convention for mass properties object to "-"}
\usage{
set_poi_conv_minus(ds, target, v)
}
\arguments{
\item{ds}{Ignored}

\item{target}{Ignored}

\item{v}{A mass properties object}
}
\value{
The mass properties object with the POI convention set to "-"
}
\description{
\code{set_poi_conv_minus()} sets the products of inertia convention for a
mass properties object to "+". This does not affect the values of
object, but it determines how products of inertia are saved to a
data set.

The signature \verb{of set_poi_conv_minus()} is such that it can be passed as an \code{override} argument
to \code{update_mass_props()} and \code{update_mass_props_and_unc()}, thus ensuring
that all calculated POI values follow the negative integral convention.
}
\examples{
set_poi_conv_minus(NULL, NULL, get_mass_props(mp_table, "C.1.2.2.3.2.1.1"))
}
