% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{set_poi_conv_plus}
\alias{set_poi_conv_plus}
\title{Set POI convention for mass properties object to "+"}
\usage{
set_poi_conv_plus(ds, target, v)
}
\arguments{
\item{ds}{Ignored.}

\item{target}{Ignored.}

\item{v}{A mass properties list.}
}
\value{
The mass properties object with the named element \code{poi_conv} set to "+"
}
\description{
\code{set_poi_conv_plus()} sets the products of inertia convention for a
mass properties list to "+". This convention determines how products of inertia are
saved to a data set.

The signature of \code{set_poi_conv_plus()} is such that it can be passed as an \code{override} argument
to \code{update_mass_props()} and \code{update_mass_props_and_unc()}, thus ensuring
that calculated POI values are saved using the positive integral convention.
}
\examples{
set_poi_conv_plus(NULL, NULL, get_mass_props(mp_table, "C.1.2.2.3.2.1.1"))

}
