% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radii.R
\name{set_radii_of_gyration}
\alias{set_radii_of_gyration}
\title{Set radii of gyration for a row in a data frame}
\usage{
set_radii_of_gyration(ds, id, rg, set_by_id = df_set_by_id)
}
\arguments{
\item{ds}{A data set with an \code{id} property.}

\item{id}{The \code{id} value of the desired item}

\item{rg}{A list with the following named elements:
\itemize{
\item \code{radii_gyration} Numeric 3x3 matrix radii of gyration.
}}

\item{set_by_id}{A function to set the value of a property for the selected item,
called as set_by_id(ds, id, property, value); default df_set_by_id()}
}
\value{
The updated data set
}
\description{
\code{set_radii_of_gyration()} sets radii of gyration for a
selected item in a data set with an \code{id} property
}
\examples{
rgl <- list(radii_gyration = c(x = 1, y = 2, z = 3))
set_radii_of_gyration(mp_table, "C.1", rgl)[1:5, ]
}
