% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_tree.R
\docType{data}
\name{test_tree}
\alias{test_tree}
\title{Example Mass Properties Tree}
\format{
An igraph tree with edges from child id to parent id.
}
\usage{
test_tree
}
\description{
Example Mass Properties Tree
}
\keyword{datasets}
