% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{validate_mass_props_and_unc}
\alias{validate_mass_props_and_unc}
\title{Validate mass properties and uncertainties}
\usage{
validate_mass_props_and_unc(mpu)
}
\arguments{
\item{mpu}{mass properties and uncertainty object}
}
\value{
TRUE if valid, stops otherwise
}
\description{
\code{validate_mass_props_and_unc()} is a convenience wrapper that calculates the logical
conjunction of \code{validate_mass_props()} and \code{validate_mass_props_unc()}.
}
\examples{
mpu <- get_mass_props_and_unc(sawe_table, "Widget")
validate_mass_props_and_unc(mpu)
}
