% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{validate_mass_props_unc}
\alias{validate_mass_props_unc}
\title{Validate mass properties uncertainties}
\usage{
validate_mass_props_unc(mp)
}
\arguments{
\item{mp}{mass properties uncertainties object}
}
\value{
TRUE if valid, stops otherwise
}
\description{
\code{validate_mass_props_and_unc()}
ensures that the following are true:
\itemize{
\item mass uncertainty is non-missing and non-negative
\item center of mass uncertainty is a 3-vector of non-missing non-negative values
\item for non-point masses, the inertia tensor uncertainty is a 3x3 matrix of non-missing non-negative values
}
}
\examples{
mp <- get_mass_props_unc(sawe_table, "Widget")
validate_mass_props_unc(mp)
}
