% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{add_radii_of_gyration}
\alias{add_radii_of_gyration}
\title{Add Radii of Gyration to a Mass Properties Dataframe}
\usage{
add_radii_of_gyration(df)
}
\arguments{
\item{df}{data frame including columns \code{mass}, \code{Ixx}, \code{Iyy}, \code{Izz}}
}
\value{
input data frame with added (or overwritten) columns \code{Kx}, \code{Ky}, \code{Kz}.
}
\description{
\code{add_radii_of_gyration()} adds radii of gyration to a dataframe with mass
and moments of inertia columns. This computation is not recursive.
}
\examples{
add_radii_of_gyration(rollup_mass_props(test_tree, test_table))
}
