% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{add_radii_of_gyration_unc}
\alias{add_radii_of_gyration_unc}
\title{Add Radii of Gyration Uncertainties}
\usage{
add_radii_of_gyration_unc(df)
}
\arguments{
\item{df}{a data frame including columns \code{mass}, 'σ_mass\verb{, }Ixx\verb{, }σ_Ixx\verb{, }Iyy\verb{, }σ_Iyy\verb{, }Izz\verb{, and }σ_Izz`.}
}
\value{
input data frame with added (or overwritten) columns \code{σ_Kx}, \code{σ_Ky}, \code{σ_Kz}.
}
\description{
\code{add_radii_of_gyration_unc()} adds radii of gyration uncertainties to a dataframe with mass
and moments of inertia and their uncertainties columns. This computation is not recursive.
}
\examples{
add_radii_of_gyration_unc(rollup_mass_props_and_unc(sawe_tree, sawe_table))
}
