% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{combine_mass_props}
\alias{combine_mass_props}
\title{Combine mass properties}
\usage{
combine_mass_props(vl)
}
\arguments{
\item{vl}{List of mass properties lists}
}
\value{
Combined mass properties list
}
\description{
Combine mass properties
}
\examples{
leaves <- test_table[which(!is.na(test_table$mass)), "id"]
vl <- Map(f = function(id) df_get_mass_props(test_table, id), leaves)
combine_mass_props(vl)
}
