% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{combine_mass_props_and_unc}
\alias{combine_mass_props_and_unc}
\title{Combine mass properties and uncertainties}
\usage{
combine_mass_props_and_unc(vl)
}
\arguments{
\item{vl}{List of mass properties and uncertainties lists}
}
\value{
Combined mass properties and uncertainties
}
\description{
Combine mass properties and uncertainties
}
\examples{
vl <- Map(f = function(id) df_get_mass_props_and_unc(sawe_table, id), list("Widget", "2nd Part"))
combine_mass_props_and_unc(vl)
}
