% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{get_mass_props}
\alias{get_mass_props}
\title{Get mass properties for a row in a data frame}
\usage{
get_mass_props(df, id)
}
\arguments{
\item{df}{A data frame with (at least) these columns: \code{id}, \code{mass}, \code{Cx},
\code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy}, \code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}.}

\item{id}{The \code{id} value of the desired row.}
}
\value{
A list with the following named elements:
\itemize{
\item \code{mass} Mass (numeric).
\item \code{center_mass} Center of mass (3-dimensional numeric vector).
\item \code{inertia} Inertia tensor (3x3 numeric matrix).
\item \code{point} Logical indicating point mass. The intertia of point masses is not included in calculations.
}
}
\description{
\code{get_mass_props()} creates a mass properties list from a specified row in a data frame.
}
\examples{
get_mass_props(mp_table, "C.1.2.2.3.1.2.3")

}
