% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massProps.R
\name{get_mass_props}
\alias{get_mass_props}
\title{Get mass properties for a row in a data frame}
\usage{
get_mass_props(df, id)
}
\arguments{
\item{df}{A data frame}

\item{id}{ID value of the desired row}
}
\value{
A list with the following named elements:
\itemize{
\item \code{mass} mass (numeric)
\item \code{center_mass} center of mass (3-dimensional numeric)
\item \code{inertia} Inertia tensor (3x3 numeric matrix)
\item \code{point} Logical indicating point mass, i.e., negligible inertia
}
}
\description{
\code{get_mass_props()} gets mass properties for a specified row in a data frame
with (at least) these columns: \code{id}, \code{mass}, \code{Cx}, \code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy},
\code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}.
}
\examples{
get_mass_props(mp_table, "C.1.2.2.3.1.2.3")
}
