% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{get_mass_props_and_unc}
\alias{get_mass_props_and_unc}
\title{Get mass properties and uncertainties for a row in a data frame}
\usage{
get_mass_props_and_unc(df, id)
}
\arguments{
\item{df}{A data frame}

\item{id}{ID value of the desired row}
}
\value{
A list with the following named elements:
\itemize{
\item \code{mass} mass (numeric)
\item \code{center_mass} center of mass (3-dimensional numeric)
\item \code{inertia} Inertia tensor (3x3 numeric matrix)
\item \code{point} Logical indicating point mass, i.e., negligible inertia
\item \code{sigma_mass} mass uncertainty
\item \code{sigma_center_mass} center of mass uncertainty (3-dimensional numeric)
\item \code{sigma_inertia} Inertia tensor uncertainty (3x3 numeric matrix)
}
}
\description{
\code{get_mass_props_and_unc()} gets mass properties with uncertainties for a specified row in a data frame
with (at least) these columns: \code{id}, \code{mass}, \code{Cx}, \code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy},
\code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}, \code{σ_mass}, \code{σ_Cx}, \code{σ_Cy}, \code{σ_Cz}, \code{σ_Ixx}, \code{σ_Iyy}, \code{σ_Izz}, \code{σ_Ixy}, \code{σ_Ixz}, \code{σ_Iyz}.
}
\examples{
get_mass_props_and_unc(mp_table, "C.1.2.2.3.1.2.3")
}
