% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops-package.R
\docType{package}
\name{massprops-package}
\alias{massprops}
\alias{massprops-package}
\title{massprops: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
