% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops-package.R
\docType{package}
\name{massProps-package}
\alias{massProps}
\alias{massProps-package}
\title{massProps: Compute Mass Properties and Uncertainties of Tree Structures}
\description{
The 'massProps' package extends 'rollupTree' with functions to recursively calculate mass properties (and optionally, their uncertainties) for arbitrary decomposition trees. R. L. Zimmerman, J. H. Nakai. (2005) \url{https://www.sawe.org/product/paper-3360/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jsjuni.github.io/massProps/}
  \item \url{https://github.com/jsjuni/massProps}
  \item Report bugs at \url{https://github.com/jsjuni/massProps/issues}
}

}
\author{
\strong{Maintainer}: James Steven Jenkins \email{sjenkins@studioj.us} (\href{https://orcid.org/0000-0002-0725-0884}{ORCID}) [copyright holder]

}
\keyword{internal}
