% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops-package.R
\docType{package}
\name{massProps-package}
\alias{massProps}
\alias{massProps-package}
\title{massProps: Computing Mass Properties of Aggregates}
\description{
The 'massProps' package calculates mass properties and uncertainties for tree structures (e.g., assembly bill of materials).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jsjuni.github.io/massProps/}
  \item \url{https://github.com/jsjuni/massProps}
  \item Report bugs at \url{https://github.com/jsjuni/massProps/issues}
}

}
\author{
\strong{Maintainer}: James Steven Jenkins \email{sjenkins@studioj.us} (\href{https://orcid.org/0000-0002-0725-0884}{ORCID}) [copyright holder]

}
\keyword{internal}
