% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tree.R
\docType{data}
\name{mp_tree}
\alias{mp_tree}
\title{Example Mass Properties Tree}
\format{
An 'igraph' tree whose vertices are named as the values of the \code{id}
column of a mass properties table and whose directed edges point from child
id to parent id.
}
\usage{
mp_tree
}
\description{
Example Mass Properties Tree
}
\keyword{datasets}
