% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{rollup_mass_props_and_unc}
\alias{rollup_mass_props_and_unc}
\title{Roll Up Mass Properties and Uncertainties}
\usage{
rollup_mass_props_and_unc(
  tree,
  df,
  validate_df = validate_mass_props_and_unc_table,
  ...
)
}
\arguments{
\item{tree}{An igraph directed graph that is a valid single-rooted in-tree with edges from child vertex to parent vertex.}

\item{df}{A mass properties and uncertainties table}

\item{validate_df}{A validator for the tree and table, default \code{validate_mass_props_and_unc_table()}}

\item{...}{Other parameters passed \code{rollup()}}
}
\value{
The updated data frame
}
\description{
'rollup_mass_props()' rolls up mass properties in a data frame
with (at least) these columns: \code{id}, \code{mass}, \code{Cx}, \code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy},
\code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}, \code{σ_mass}, \code{σ_Cx}, \code{σ_Cy}, \code{σ_Cz}, \code{σ_Ixx}, \code{σ_Iyy}, \code{σ_Izz}, \code{σ_Ixy}, \code{σ_Ixz}, \code{σ_Iyz}.
}
\examples{
rollup_mass_props_and_unc(sawe_tree, sawe_table)
}
