% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{rollup_mass_props_unc_fast}
\alias{rollup_mass_props_unc_fast}
\title{Roll Up Mass Properties Uncertainties Without Input Validation}
\usage{
rollup_mass_props_unc_fast(tree, df)
}
\arguments{
\item{tree}{tree passed to \code{rollup()}}

\item{df}{mass properties and uncertainties table passed to \code{rollup()}}
}
\value{
The updated data frame
}
\description{
\code{rollup_mass_props_unc_fast()} performs the same operation as \code{rollup_mass_props_unc()}
but omits input validation. It is somewhat faster than  \code{rollup_mass_props_unc()} but should
be used with caution and only under circumstances in which the caller assumes
responsibility for validity of input. Its behavior when passed ill-formed input is unspecified.
}
\examples{
rollup_mass_props_unc_fast(sawe_tree, sawe_table)
}
