% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sawe_tree.R
\docType{data}
\name{sawe_tree}
\alias{sawe_tree}
\title{Mass Properties and Uncertainties Tree from SAWE Paper No. 3360}
\format{
An igraph tree with edges from child id to parent id.
}
\source{
Zimmerman, Robert L., and John H. Nakai. 2005. “Are You Sure?
Uncertainty in Mass Properties Engineering.” In 64th Annual International
Conference on Mass Properties Engineering, 123–60. Society of Allied Weight
Engineers.
}
\usage{
sawe_tree
}
\description{
Mass Properties and Uncertainties Tree from SAWE Paper No. 3360
}
\keyword{datasets}
