% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massProps.R
\name{set_mass_props}
\alias{set_mass_props}
\title{Set mass properties for a row in a data frame}
\usage{
set_mass_props(df, id, v)
}
\arguments{
\item{df}{A data frame}

\item{id}{ID value of the desired row}

\item{v}{#' A list with the following named elements:
\itemize{
\item \code{mass} mass (numeric)
\item \code{center_mass} center of mass (3-dimensional numeric)
\item \code{inertia} Inertia tensor (3x3 numeric matrix)
\item \code{point} Logical indicating point mass, i.e., negligible inertia
\item \code{poi_conv} Enumeration c("+", "-") indicating sign convention for products of inertia
}}
}
\value{
The updated data frame
}
\description{
\code{set_mass_props()} sets mass properties for a specified row in a data frame with
an \code{id} column.
}
\examples{
df <- data.frame(id = c("C.1.2.2.3.1.2.3", "C.1.2.2.3.2.1.1"))
v <- get_mass_props(mp_table, "C.1.2.2.3.2.1.1")
v$poi_conv = "+"
set_mass_props(df, "C.1.2.2.3.2.1.1", v)
}
