% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{set_mass_props_and_unc}
\alias{set_mass_props_and_unc}
\title{Set mass properties and uncertainties for a row in a data frame}
\usage{
set_mass_props_and_unc(df, id, v)
}
\arguments{
\item{df}{A data frame with an \code{id} column.}

\item{id}{The \code{id} value of the desired row.}

\item{v}{A list containing the following named elements:
\itemize{
\item \code{mass} mass (numeric)
\item \code{center_mass} center of mass (3-dimensional numeric)
\item \code{inertia} Inertia tensor (3x3 numeric matrix)
\item \code{point} Logical indicating point mass, i.e., negligible inertia
\item \code{poi_conv} Enumeration c("+", "-") indicating sign convention for products of inertia
\item \code{sigma_mass} mass uncertainty
\item \code{sigma_center_mass} center of mass uncertainty (3-dimensional numeric)
\item \code{sigma_inertia} Inertia tensor uncertainty (3x3 numeric matrix)
}}
}
\value{
The updated data frame.
}
\description{
\code{set_mass_props_and_unc()} is a convenience wrapper that combines the results of
\code{set_mass_props()} and \code{set_mass_props_unc()}.
}
\examples{
mpu <- c(get_mass_props_and_unc(sawe_table, "Widget"), poi_conv = "+")
set_mass_props_and_unc(sawe_table, "Combined", mpu)
}
