% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{set_mass_props_unc}
\alias{set_mass_props_unc}
\title{Set mass properties uncertainties for a row in a data frame}
\usage{
set_mass_props_unc(df, id, v)
}
\arguments{
\item{df}{A data frame}

\item{id}{ID value of the desired row}

\item{v}{#' A list with the following named elements:
\itemize{
\item \code{sigma_mass} mass uncertainty (numeric)
\item \code{sigma_center_mass} center of mass uncertainty (3-dimensional numeric)
\item \code{sigma_inertia} Inertia tensor uncertainty (3x3 numeric matrix)
}}
}
\value{
The updated data frame
}
\description{
\code{set_mass_props_unc()} sets mass properties and uncertainties for a
specified row in a data frame with an \code{id} column.
}
\examples{
set_mass_props_unc(sawe_table, "Combined", get_mass_props_unc(sawe_table, "Widget"))
}
