% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{update_mass_props}
\alias{update_mass_props}
\title{Update mass properties}
\usage{
update_mass_props(df, target, sources, override = set_poi_conv_from_target)
}
\arguments{
\item{df}{data frame}

\item{target}{id of the target row}

\item{sources}{ids of the source rows}

\item{override}{override function, called as override(df, target, value)}
}
\value{
updated data frame
}
\description{
\code{update_mass_props()} updates mass properties for a specified target row from
specified source rows in a data frame
with (at least) these columns: \code{id}, \code{mass}, \code{Cx}, \code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy},
\code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}.
}
\examples{
leaves <- names(igraph::neighbors(test_tree, "A.3", mode = "in"))
df <- update_mass_props(test_table, "A.3", leaves)
get_mass_props(df, "A.3")
}
