% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{update_mass_props_and_unc}
\alias{update_mass_props_and_unc}
\title{Update mass properties and uncertainties}
\usage{
update_mass_props_and_unc(
  df,
  target,
  sources,
  override = set_poi_conv_from_target
)
}
\arguments{
\item{df}{A data frame}

\item{target}{ID of the target row}

\item{sources}{IDs of the source rows}

\item{override}{An override function, called as override(df, target, value)}
}
\value{
The updated data frame
}
\description{
\code{update_mass_props_and_unc()} updates mass properties and uncertainties
for a specified target row from
specified source rows in a data frame
with (at least) these columns: \code{id}, \code{mass}, \code{Cx}, \code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy},
\code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}, \code{σ_mass}, \code{σ_Cx}, \code{σ_Cy}, \code{σ_Cz}, \code{σ_Ixx}, \code{σ_Iyy}, \code{σ_Izz}, \code{σ_Ixy}, \code{σ_Ixz}, \code{σ_Iyz}.
}
\examples{
leaves <- list("Widget", "2nd Part")
df <- update_mass_props_and_unc(sawe_table, "Combined", leaves)
}
