% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{update_mass_props_unc}
\alias{update_mass_props_unc}
\title{Update mass properties}
\usage{
update_mass_props_unc(df, target, sources, override = set_poi_conv_from_target)
}
\arguments{
\item{df}{data frame}

\item{target}{id of the target row}

\item{sources}{ids of the source rows}

\item{override}{override function, called as override(df, target, value)}
}
\value{
updated data frame
}
\description{
\code{update_mass_props_unc()} updates mass properties uncertainties
for a specified target row from
specified source rows in a data frame
with (at least) these columns: \code{id}, \code{sigma_mass}, \code{sigma_Cx}, \code{sigma_Cy}, \code{sigma_Cz},
\code{sigma_Ixx}, \code{sigma_Iyy}, \code{sigma_Izz}, \code{sigma_Ixy}, \code{sigma_Ixz}, \code{sigma_Iyz}.
}
\examples{
leaves <- names(igraph::neighbors(sawe_tree, "Combined", mode = "in"))
df <- update_mass_props_unc(sawe_table, "Combined", leaves)
get_mass_props_unc(df, "Combined")
}
