% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massProps.R
\name{validate_mass_props}
\alias{validate_mass_props}
\title{Validate mass properties}
\usage{
validate_mass_props(mp)
}
\arguments{
\item{mp}{Mass properties object}
}
\value{
TRUE if valid, stops otherwise
}
\description{
\code{validate_mass_props()} ensures that a mass properties object satisfies the following
constraints:
\itemize{
\item mass is non-missing and positive
\item center of mass is a 3-vector of non-missing numeric values
\item point mass indicator is TRUE or FALSE
\item for every non-point mass:
\itemize{
\item the inertia tensor is positive definite
\item eigenvalues of the inertia tensor satisfy the triangle inequalities:
\itemize{
\item e1 < e2 + e3
\item e2 < e1 + e3
\item e3 < e1 + e2
}
}
}
}
\examples{
mp <- get_mass_props(test_table, "C.1")
validate_mass_props(mp)
}
