% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{validate_mass_props_and_unc_table}
\alias{validate_mass_props_and_unc_table}
\title{Validate a mass properties and uncertainties table}
\usage{
validate_mass_props_and_unc_table(
  tree,
  ds,
  get_ids = df_get_ids,
  get_mpu = get_mass_props_and_unc,
  validate = validate_mass_props_and_unc
)
}
\arguments{
\item{tree}{An 'igraph' tree whose vertices are named as the values of the \code{id}
column of \code{ds} and whose directed edges point from child id to parent id.}

\item{ds}{A data set  with (at least) these properties: \code{id}, \code{mass}, \code{Cx},
\code{Cy}, \code{Cz}, \code{Ixx}, \code{Iyy}, \code{Izz}, \code{Ixy}, \code{Ixz}, \code{Iyz}, \code{POIconv}, \code{Ipoint}.}

\item{get_ids}{A function to get all ids from the data set, called as get_ids(ds). Default: df_get_ids().}

\item{get_mpu}{A function get the mass properties and uncertainties for a selected item, called as get_mpu(ds, id). Default: get_mass_props_and_unc().}

\item{validate}{A function to validate a set of mass properties, called as validate(mp). Default: validate_mass_props_and_unc().}
}
\value{
TRUE if valid, stops with an error otherwise
}
\description{
\code{validate_mass_props_and_unc()} calls
\code{validate_mass_props_table()} and further applies the checks of
\code{validate_mass_props_and_unc()} to every item of the data set corresponding
to a leaf vertex of the tree.
}
\examples{
validate_mass_props_and_unc_table(mp_tree_small, mp_table_small)
}
