% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massprops.R
\name{validate_mass_props_and_unc_table}
\alias{validate_mass_props_and_unc_table}
\title{Validate a mass properties and uncertainties table}
\usage{
validate_mass_props_and_unc_table(tree, df)
}
\arguments{
\item{tree}{An \code{igraph} tree with edges from child to parent}

\item{df}{A data frame to validate}
}
\value{
TRUE if valid, stops with an error otherwise
}
\description{
\code{validate_mass_props_and_unc()} calls
\code{validate_mass_props_table()} and further ensures that the mass properties uncertainties of every leaf element
of the table are valid.
}
\examples{
validate_mass_props_and_unc_table(mp_tree, mp_table)
}
