% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massProps.R
\name{validate_mass_props_table}
\alias{validate_mass_props_table}
\title{Validate a mass properties table}
\usage{
validate_mass_props_table(tree, df)
}
\arguments{
\item{tree}{An \code{igraph} tree with edges from child to parent}

\item{df}{A data frame to validate}
}
\value{
TRUE if valid, stops with an error otherwise
}
\description{
\code{validate_mass_props_table()} ensures that the \code{id} column of the table and the vertices
of the tree contain the same identifiers, and that the mass properties of every leaf element
of the table are valid.
}
\examples{
validate_mass_props_table(mp_tree, mp_table)
}
